Move common methods to utils
