+++ test/distributed/org/apache/cassandra/distributed/test/auth/MutualTlsCertificateValidityPeriodTest.java
@@ -18,9 +18,7 @@
 
 package org.apache.cassandra.distributed.test.auth;
 
-import java.io.File;
 import java.net.InetAddress;
-import java.net.InetSocketAddress;
 import java.nio.file.Path;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
@@ -28,11 +26,8 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Consumer;
-import java.util.function.Function;
 import java.util.stream.StreamSupport;
-import javax.net.ssl.SSLException;
 
-import com.google.common.collect.ImmutableMap;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -41,38 +36,29 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 import com.codahale.metrics.Histogram;
-import com.datastax.driver.core.PlainTextAuthProvider;
-import com.datastax.driver.core.RemoteEndpointAwareJdkSSLOptions;
 import com.datastax.driver.core.ResultSet;
 import com.datastax.driver.core.Row;
-import com.datastax.driver.core.SSLOptions;
 import com.datastax.driver.core.Session;
 import com.datastax.driver.core.SimpleStatement;
-import com.datastax.driver.core.policies.LoadBalancingPolicy;
-import org.apache.cassandra.config.EncryptionOptions;
 import org.apache.cassandra.distributed.Cluster;
 import org.apache.cassandra.distributed.api.Feature;
 import org.apache.cassandra.distributed.api.ICluster;
 import org.apache.cassandra.distributed.api.IInvokableInstance;
-import org.apache.cassandra.distributed.shared.ClusterUtils;
 import org.apache.cassandra.distributed.test.JavaDriverUtils;
 import org.apache.cassandra.distributed.test.TestBaseImpl;
-import org.apache.cassandra.distributed.util.Auth;
-import org.apache.cassandra.distributed.util.SingleHostLoadBalancingPolicy;
 import org.apache.cassandra.metrics.ClearableHistogram;
 import org.apache.cassandra.metrics.MutualTlsMetrics;
-import org.apache.cassandra.security.ISslContextFactory;
-import org.apache.cassandra.transport.SimpleClientSslContextFactory;
 import org.apache.cassandra.utils.tls.CertificateBuilder;
 import org.apache.cassandra.utils.tls.CertificateBundle;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.InstanceOfAssertFactories;
 
-import static org.apache.cassandra.auth.CassandraRoleManager.DEFAULT_SUPERUSER_NAME;
-import static org.apache.cassandra.auth.CassandraRoleManager.DEFAULT_SUPERUSER_PASSWORD;
-import static org.apache.cassandra.transport.TlsTestUtils.CLIENT_TRUSTSTORE_PASSWORD;
+import static org.apache.cassandra.transport.TlsTestUtils.CLIENT_SPIFFE_IDENTITY;
 import static org.apache.cassandra.transport.TlsTestUtils.SERVER_KEYSTORE_PASSWORD;
 import static org.apache.cassandra.transport.TlsTestUtils.SERVER_TRUSTSTORE_PASSWORD;
+import static org.apache.cassandra.transport.TlsTestUtils.configureIdentity;
+import static org.apache.cassandra.transport.TlsTestUtils.generateClientCertificate;
+import static org.apache.cassandra.transport.TlsTestUtils.getSSLOptions;
 import static org.assertj.core.api.Assertions.as;
 import static org.assertj.core.api.Assertions.fail;
 
@@ -81,9 +67,7 @@ import static org.assertj.core.api.Assertions.fail;
  */
 public class MutualTlsCertificateValidityPeriodTest extends TestBaseImpl
 {
-    private static final String IDENTITY = "spiffe://test.cassandra.apache.org/dTest/mtls";
     private static ICluster<IInvokableInstance> CLUSTER;
-    private static final char[] KEYSTORE_PASSWORD = "cassandra".toCharArray();
 
     @ClassRule
     public static TemporaryFolder tempFolder = new TemporaryFolder();
@@ -125,10 +109,9 @@ public class MutualTlsCertificateValidityPeriodTest extends TestBaseImpl
                                  .set("client_encryption_options.keystore_password", SERVER_KEYSTORE_PASSWORD)
                                  .set("client_encryption_options.truststore", truststorePath.toString())
                                  .set("client_encryption_options.truststore_password", SERVER_TRUSTSTORE_PASSWORD)
-                                 .set("client_encryption_options.require_endpoint_verification", "true")
-                                 .set("audit_logging_options.enabled", "true")
-                                 .set("audit_logging_options.logger.class_name", "FileAuditLogger")
-                                 .set("audit_logging_options.included_categories", "AUTH")
+                                 .set("client_encryption_options.require_endpoint_verification", "false")
+                                 .set("client_encryption_options.max_certificate_validity_period", "30d")
+                                 .set("client_encryption_options.certificate_validity_warn_threshold", "5d")
                                  .with(Feature.NATIVE_PROTOCOL, Feature.GOSSIP));
         CLUSTER = builder.start();
 
@@ -183,9 +166,9 @@ public class MutualTlsCertificateValidityPeriodTest extends TestBaseImpl
 
             Assertions.assertThat(authenticationMetadata).isNotNull().hasSize(1)
                       .containsKey("identity")
-                      .extractingByKey("identity", as(InstanceOfAssertFactories.STRING)).isEqualTo(IDENTITY);
+                      .extractingByKey("identity", as(InstanceOfAssertFactories.STRING)).isEqualTo(CLIENT_SPIFFE_IDENTITY);
             Assertions.assertThat(row.getString("authentication_mode")).isEqualTo("MutualTls");
-            Assertions.assertThat(CLUSTER.get(1).logs().grep("Certificate with identity '" + IDENTITY + "' will expire").getResult())
+            Assertions.assertThat(CLUSTER.get(1).logs().grep("Certificate with identity '" + CLIENT_SPIFFE_IDENTITY + "' will expire").getResult())
                       .isNotEmpty();
             CLUSTER.get(1).runOnInstance(() -> Assertions.assertThat(MutualTlsMetrics.instance.clientCertificateExpirationDays.getCount()).isEqualTo(2));
         });
@@ -213,9 +196,9 @@ public class MutualTlsCertificateValidityPeriodTest extends TestBaseImpl
 
             Assertions.assertThat(authenticationMetadata).isNotNull().hasSize(1)
                       .containsKey("identity")
-                      .extractingByKey("identity", as(InstanceOfAssertFactories.STRING)).isEqualTo(IDENTITY);
+                      .extractingByKey("identity", as(InstanceOfAssertFactories.STRING)).isEqualTo(CLIENT_SPIFFE_IDENTITY);
             Assertions.assertThat(row.getString("authentication_mode")).isEqualTo("MutualTls");
-            Assertions.assertThat(CLUSTER.get(1).logs().grep("Certificate with identity '" + IDENTITY + "' will expire").getResult())
+            Assertions.assertThat(CLUSTER.get(1).logs().grep("Certificate with identity '" + CLIENT_SPIFFE_IDENTITY + "' will expire").getResult())
                       .isNotEmpty();
             CLUSTER.get(1).runOnInstance(() -> Assertions.assertThat(MutualTlsMetrics.instance.clientCertificateExpirationDays.getCount()).isGreaterThanOrEqualTo(2));
         });
@@ -270,94 +253,5 @@ public class MutualTlsCertificateValidityPeriodTest extends TestBaseImpl
                 consumer.accept(session);
             }
         }
-    }
-
-    public static SSLOptions getSSLOptions(Path keystorePath, Path truststorePath) throws RuntimeException
-    {
-        try
-        {
-            return RemoteEndpointAwareJdkSSLOptions.builder()
-                                                   .withSSLContext(getClientSslContextFactory(keystorePath, truststorePath)
-                                                                   .createJSSESslContext(EncryptionOptions.ClientAuth.OPTIONAL))
-                                                   .build();
-        }
-        catch (SSLException e)
-        {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static ISslContextFactory getClientSslContextFactory(Path keystorePath, Path truststorePath)
-    {
-        ImmutableMap.Builder<String, Object> params = ImmutableMap.<String, Object>builder()
-                                                                  .put("truststore", truststorePath.toString())
-                                                                  .put("truststore_password", CLIENT_TRUSTSTORE_PASSWORD);
-
-        if (keystorePath != null)
-        {
-            params.put("keystore", keystorePath.toString())
-                  .put("keystore_password", "cassandra");
-        }
-
-        return new SimpleClientSslContextFactory(params.build());
-    }
-
-    static void configureIdentity(ICluster<IInvokableInstance> cluster, SSLOptions sslOptions)
-    {
-        withAuthenticatedSession(cluster.get(1), DEFAULT_SUPERUSER_NAME, DEFAULT_SUPERUSER_PASSWORD, session -> {
-            session.execute("CREATE ROLE cassandra_ssl_test WITH LOGIN = true");
-            session.execute(String.format("ADD IDENTITY '%s' TO ROLE 'cassandra_ssl_test'", IDENTITY));
-            // GRANT select to cassandra_ssl_test to be able to query the system_views.clients virtual table
-            session.execute("GRANT SELECT ON ALL KEYSPACES to cassandra_ssl_test");
-        }, sslOptions);
-    }
-
-    static void withAuthenticatedSession(IInvokableInstance instance,
-                                         String username,
-                                         String password,
-                                         Consumer<Session> consumer,
-                                         SSLOptions sslOptions)
-    {
-        // wait for existing roles
-        Auth.waitForExistingRoles(instance);
-
-        InetSocketAddress nativeInetSocketAddress = ClusterUtils.getNativeInetSocketAddress(instance);
-        InetAddress address = nativeInetSocketAddress.getAddress();
-        LoadBalancingPolicy lbc = new SingleHostLoadBalancingPolicy(address);
-
-        com.datastax.driver.core.Cluster.Builder builder = com.datastax.driver.core.Cluster.builder()
-                                                                                           .withLoadBalancingPolicy(lbc)
-                                                                                           .withSSL(sslOptions)
-                                                                                           .withAuthProvider(new PlainTextAuthProvider(username, password))
-                                                                                           .addContactPoint(address.getHostAddress())
-                                                                                           .withPort(nativeInetSocketAddress.getPort());
-
-        try (com.datastax.driver.core.Cluster c = builder.build(); Session session = c.connect())
-        {
-            consumer.accept(session);
-        }
-    }
-
-    static Path generateSelfSignedCertificate(Function<CertificateBuilder, CertificateBuilder> customizeCertificate, File targetDirectory) throws Exception
-    {
-        return generateClientCertificate(customizeCertificate, targetDirectory, null);
-    }
-
-    static Path generateClientCertificate(Function<CertificateBuilder, CertificateBuilder> customizeCertificate, File targetDirectory, CertificateBundle ca) throws Exception
-    {
-        CertificateBuilder builder = new CertificateBuilder().subject("CN=Apache Cassandra, OU=ssl_test, O=Unknown, L=Unknown, ST=Unknown, C=Unknown")
-                                                             .notBefore(Instant.now().minus(1, ChronoUnit.DAYS))
-                                                             .notAfter(Instant.now().plus(1, ChronoUnit.DAYS))
-                                                             .alias("spiffecert")
-                                                             .addSanUriName(IDENTITY)
-                                                             .rsa2048Algorithm();
-        if (customizeCertificate != null)
-        {
-            builder = customizeCertificate.apply(builder);
-        }
-        CertificateBundle ssc = ca != null
-                                ? builder.buildIssuedBy(ca)
-                                : builder.buildSelfSigned();
-        return ssc.toTempKeyStorePath(targetDirectory.toPath(), KEYSTORE_PASSWORD, KEYSTORE_PASSWORD);
     }
 }
